Rename module to raw_value.rs
